Fix accessibility issues after restructuring
