+++ app/src/main/AndroidManifest.xml
@@ -9,10 +9,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".PrivacyActivity"></activity>
-        <activity android:name=".CreateActivity" />
+        <activity android:name=".activities.PrivacyActivity"></activity>
+        <activity android:name=".activities.CreateActivity" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
