+++ app/src/main/res/layout/activity_create.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_margin="16dp"
-    tools:context=".CreateActivity">
+    tools:context=".activities.CreateActivity">
 
     <TextView
         android:id="@+id/label_name"
